test(wiring): relocate WIRE-043 to a sibling file (FR-006 net-shrink)
